Angular 6 upgrade

See merge request kohese/kohese-server!280

+++ kohese-portal/client/src/services/item-repository/item-repository.service.ts
@@ -1,5 +1,6 @@
-import { Injectable } from '@angular/core';
 
+import { map} from 'rxjs/operators';
+import { Injectable } from '@angular/core';
 import * as _ from 'underscore';
 
 import { SocketService } from '../socket/socket.service';
@@ -14,11 +15,8 @@ import { ItemCache } from '../../../../common/src/item-cache';
 import { ItemProxy } from '../../../../common/src/item-proxy';
 import { KoheseModel } from '../../../../common/src/KoheseModel';
 
-import { Subject } from 'rxjs/Subject';
-import { BehaviorSubject } from 'rxjs/BehaviorSubject';
-import { Subscription } from 'rxjs/Subscription';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
+import { Subject ,  BehaviorSubject ,  Subscription ,  Observable, bindCallback } from 'rxjs';
+
 import { LogService } from '../log/log.service';
 import { InitializeLogs } from './item-repository.registry';
 
@@ -114,7 +112,7 @@ export class ItemRepository {
         for (let idx in script.attributes) {
           let attribute: any = script.attributes[idx];
           if (attribute.value) {
-            if (attribute.value.match(/^cache-worker/)) {
+            if (attribute.value.match(/^scripts/)) {
               cacheWorkerBundle = attribute.value;
               break scriptLoop;
             }
@@ -528,7 +526,9 @@ export class ItemRepository {
     if (response.deleteItems) {
       response.deleteItems.forEach((deletedItemId) => {
         var proxy = ItemProxy.getWorkingTree().getProxyFor(deletedItemId);
-        proxy.deleteItem();
+        if (proxy) {
+          proxy.deleteItem();
+        }
       });
     }
   }
@@ -673,15 +673,15 @@ export class ItemRepository {
   //////////////////////////////////////////////////////////////////////////
   public getHistoryFor(proxy: ItemProxy): Observable<Array<any>> {
     let emitReturningObservable: (message: string, data: any) => Observable<any> =
-      Observable.bindCallback(this.socketService.getSocket().emit.bind(this.
+      bindCallback(this.socketService.getSocket().emit.bind(this.
         socketService.getSocket()));
-    return emitReturningObservable('Item/getHistory', { onId: proxy.item.id }).
+    return emitReturningObservable('Item/getHistory', { onId: proxy.item.id }).pipe(
       map((response: any) => {
         proxy.history = response.history;
         /* Return a copy of the history so that subscribers may modify the
         returned history, if desired. */
         return JSON.parse(JSON.stringify(proxy.history));
-      });
+      }));
   }
 
   //////////////////////////////////////////////////////////////////////////
